Add fit-to-view
